requirements.txt
aiogram>=3.10,<4.0
aiohttp>=3.9,<4.0
aiosqlite>=0.20
python-dotenv>=1.0
